definitions/requirements.txt
bitproto==1.2.1
pyyaml==6.0.1
jsonschema==4.25.1